Extract word collector to a class
